requirements.txt
pytorch
scikit-learn
numpy
matplotlib
pandas
seaborn
pytorch torchvision